Rename test_cli.py for consistency
