+++ pysr/test/__init__.py
@@ -2,4 +2,4 @@ from .test import runtests
 from .test_env import runtests as runtests_env
 from .test_jax import runtests as runtests_jax
 from .test_torch import runtests as runtests_torch
-from .cliTest import runtests as runtests_cli
+from .test_cli import runtests as runtests_cli
