+++ pysr/test/test_cli.py
